pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "simple-ui-client"
version = "0.1.0"
description = "A modern, cross-platform background agent for Simple UI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Simple UI Team" }
]
keywords = ["simple-ui", "agent", "socketio", "document-conversion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "typer>=0.9.0",
    "python-socketio[asyncio_client]>=5.10.0",
    "pydantic-settings>=2.1.0",
    "pydantic>=2.5.0",
    "loguru>=0.7.2",
    "rich>=13.7.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "pywin32>=306; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
simple-client = "simple_ui_client.cli.commands:app"

[project.urls]
Homepage = "https://github.com/simple-ui/simple-ui-client"
Documentation = "https://github.com/simple-ui/simple-ui-client#readme"
Repository = "https://github.com/simple-ui/simple-ui-client"

[tool.hatch.build.targets.wheel]
packages = ["src/simple_ui_client"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
